feat: add authorization to all controllers

+++ src/main/java/org/spine/iquestionapi/controller/EntryController.java
@@ -1,7 +1,11 @@
 package org.spine.iquestionapi.controller;
 
+import java.util.List;
+
+import org.spine.iquestionapi.model.User;
 import org.spine.iquestionapi.model.Entry;
 import org.spine.iquestionapi.repository.EntryRepo;
+import org.spine.iquestionapi.service.AuthorizationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -12,25 +16,48 @@ import org.springframework.web.server.ResponseStatusException;
 @ResponseStatus(HttpStatus.OK)
 public class EntryController {
     @Autowired private EntryRepo entryRepo;
+    @Autowired private AuthorizationService authorizationService;
 
     // Get entry by id
     @GetMapping("/{id}")
     @ResponseBody
     public Entry getEntryById(@PathVariable(value="id") long id){
-        return entryRepo.findById(id).get();
+        List<Entry> entries = authorizationService.getLoggedInUser().getEntries();
+
+        for (Entry entry : entries) {
+            if (entry.getId() == id) {
+                return entry;
+            }
+        }
+
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to view this entry.");
     }
 
     // Delete an entry
     @DeleteMapping("/{id}")
     @ResponseBody
     public void deleteEntry(@PathVariable(value="id") long id){
-        entryRepo.deleteById(id);
+        List<Entry> entries = authorizationService.getLoggedInUser().getEntries();
+
+        for (Entry entry : entries) {
+            if (entry.getId() == id) {
+                entryRepo.deleteById(id);
+                return;
+            }
+        }
+
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to view this entry.");
     }
 
     // Create an entry
     @PutMapping("/")
     @ResponseBody
     public Entry createEntry(@RequestBody Entry entry){
+        // Check if user is a caregiver
+        if (authorizationService.getLoggedInUser().getRole() != User.Role.CAREGIVER) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not a caregiver.");
+        }
+
         return entryRepo.save(entry);
     }
 }
