+++ src/main/java/org/spine/iquestionapi/controller/QuestionnaireController.java
@@ -1,7 +1,9 @@
 package org.spine.iquestionapi.controller;
 
+import org.spine.iquestionapi.model.User;
 import org.spine.iquestionapi.model.Questionnaire;
 import org.spine.iquestionapi.repository.QuestionnaireRepo;
+import org.spine.iquestionapi.service.AuthorizationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -13,11 +15,17 @@ import org.springframework.web.server.ResponseStatusException;
 public class QuestionnaireController {
 
     @Autowired private QuestionnaireRepo questionnaireRepo;
+    @Autowired private AuthorizationService authorizationService;
 
     // Get all questionnaires
     @GetMapping("/all")
 
     public Questionnaire[] getAllQuestionnaires(){
+        // Check if logged in user is spine user
+        if (authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_USER || authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not an employee.");
+        }
+
         return questionnaireRepo.findAll().toArray(new Questionnaire[0]);
     }
 
@@ -25,6 +33,11 @@ public class QuestionnaireController {
     @GetMapping("/{id}")
     @ResponseBody
     public Questionnaire getQuestionnaireById(@PathVariable(value="id") long id){
+        // Check if logged in user is spine user
+        if (authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_USER || authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not an employee.");
+        }
+
         return questionnaireRepo.findById(id).get();
     }
 
@@ -32,6 +45,11 @@ public class QuestionnaireController {
     @PutMapping("/")
     @ResponseBody
     public Questionnaire createQuestionnaire(@RequestBody Questionnaire questionnaire){
+        // Check if logged in user is spine user
+        if (authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_USER || authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not an employee.");
+        }
+
         return questionnaireRepo.save(questionnaire);
     }
 
@@ -39,6 +57,11 @@ public class QuestionnaireController {
     @DeleteMapping("/{id}")
     @ResponseBody
     public void deleteQuestionnaire(@PathVariable(value="id") long id){
+        // Check if logged in user is spine user
+        if (authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_USER || authorizationService.getLoggedInUser().getRole() != User.Role.SPINE_ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not an employee.");
+        }
+        
         questionnaireRepo.deleteById(id);
     }
